requirements.txt
Django==2.2.9
djangorestframework==3.11.1
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
